sửa lại tính năng bình luận

+++ controllers/news.controller.js
@@ -1,5 +1,7 @@
 const newsModel = require('../models/news.model');
 const commentModel = require('../models/comment.model');
+
+const DeleteComment = 'You can not delete this comment';
 exports.getAllNews = async (req, res) => {
     const news = await newsModel.find({});
     res.render('./news/news', {news: news});
@@ -19,54 +21,24 @@ exports.postAddNews = async (req, res) => {
     });
     res.redirect('/news');
 }
+
+let newId;
+let comments;
 exports.getNewsById = async (req, res) => {
     console.log('teaat' + req.session);
-    const news = await newsModel.findById(req.params.id);
-    const comments = await commentModel.find({idNews: req.params.id});
+    newId = await newsModel.findById(req.params.id);
+    comments = await commentModel.find({idNews: req.params.id});
     if (!req.session.user) {
         return res.redirect('/login');
     }
-    let comment = []
-    await comments.forEach(function (item) {
-        if (item.Email === req.session.user.Email) {
-            comment.push({
-                _id: item._id,
-                idNews: item.idNews,
-                fullNameUser: 'Bạn',
-                Email: item.Email,
-                content: item.content,
-                date: item.date,
-                modelDeleteComment: true,
-            })
-        } else if (req.session.user.quyen === 'Admin') {
-            comment.push({
-                _id: item._id,
-                idNews: item.idNews,
-                fullNameUser: item.fullNameUser,
-                Email: item.Email,
-                content: item.content,
-                date: item.date,
-                modelDeleteComment: true,
-            })
-        } else {
-            comment.push({
-                _id: item._id,
-                idNews: item.idNews,
-                fullNameUser: item.fullNameUser,
-                Email: item.Email,
-                content: item.content,
-                date: item.date,
-            })
-        }
-    });
-    res.render('./news/postNews', {news: news, comments: comment, user: req.session.user});
+    res.render('./news/postNews', {news: newId, comments: comments, user: req.session.user});
 }
 exports.postEditNews = async (req, res) => {
-    const news = await newsModel.findByIdAndUpdate(req.params.id, req.body);
+    await newsModel.findByIdAndUpdate(req.params.id, req.body);
     res.redirect('/news');
 }
 exports.deleteNews = async (req, res) => {
-    const news = await newsModel.findByIdAndDelete(req.params.id);
+    await newsModel.findByIdAndDelete(req.params.id);
     res.redirect('/news');
 }
 exports.postComment = async (req, res) => {
@@ -90,7 +62,14 @@ exports.postComment = async (req, res) => {
         res.redirect('/login');
 }
 exports.postDeleteComment = async (req, res) => {
-    const comment = await commentModel.findByIdAndDelete(req.body.deleteIdComment);
+    const comment = await commentModel.findById(req.body.deleteIdComment);
     console.log(comment);
-    res.redirect('/news/id/' + comment.idNews);
+    if (comment.Email === req.session.user.Email || req.session.user.quyen === 'Admin') {
+        await commentModel.findByIdAndDelete(req.body.deleteIdComment);
+        comments = await commentModel.find({idNews: comment.idNews});
+        res.render('./news/postNews', {news: newId, comments: comments, user: req.session.user});
+    }
+    else {
+        res.render('./news/postNews', {news: newId, comments: comments, user: req.session.user, DeleteComment: DeleteComment});
+    }
 }
